requirements.txt
streamlit>=1.28.0
langchain-google-genai>=1.0.0
langchain>=0.1.0
python-dotenv>=1.0.0
streamlit-lottie>=0.0.5
streamlit-option-menu>=0.3.6
streamlit-card>=0.0.5
requests>=2.31.0
typing-extensions>=4.8.0